Remove MoveAssayRunsAction and move implementation to ExpRunTableImpl.RunTableUpdateService moveRows()
