+++ assay/src/org/labkey/assay/AssayController.java
@@ -185,7 +185,6 @@ public class AssayController extends SpringActionController
         GetProtocolAction.class,
         ImportAction.class,
         ImportRunApiAction.class,
-        MoveAssayRunsAction.class,
         PipelineDataCollectorRedirectAction.class,
         ReimportRedirectAction.class,
         SaveAssayBatchAction.class,
